requirements-conda.txt
numpy
matplotlib
biopython
cutadapt
nose
mock
pyyaml
sphinx
sphinx_rtd_theme
sh
docopt
# Non python packages
bwa==0.7.12 # Previous version was 0.7.6a from setup.py
samtools==0.1.19_ccf1da91b # See issue #202
trimmomatic==0.35 # Previous version was 0.32 from setup.py
imagemagick==6.9.3
igv==2.3.65
